+++ CMakeLists.txt
@@ -41,7 +41,7 @@ FetchContent_MakeAvailable(cmake_git_version_tracking)
 add_library(
     NUTC-client_lib OBJECT
     src/rabbitmq/rabbitmq.cpp
-    src/firebase/fetching.cpp
+    src/firebase/firebase.cpp
     src/pywrapper/runtime.cpp
     # Utils
     src/logging.cpp
